+++ src/compas_view2/app/controller.py
@@ -3,13 +3,13 @@ from OpenGL import GL
 from qtpy import QtWidgets
 from qtpy import QtCore
 
-from typing import Union
+from typing import Union, Dict
 from pathlib import Path
 
 from compas.geometry import Point
 from compas.datastructures import Network
 from compas.datastructures import Mesh
-
+from compas_view2.scene import Mouse
 from compas_view2.forms import AddForm
 from compas_view2.forms import PropertyForm
 
@@ -26,8 +26,128 @@ class Controller:
 
     """
 
-    def __init__(self, app):
+    def __init__(self, app, controller_config: Dict):
         self.app = app
+        self.config = controller_config
+        self.mouse_key = controller_config["actions"]["mouse_key"]
+        self.keys = controller_config["actions"]["keys"]
+        self.mouse = Mouse()
+        self.key_status = {}
+        self.actions = {}  # Currently, no customized actions supported.
+        self.supported_keys = {
+            "shift": QtCore.Qt.Key_Shift,
+            "control": QtCore.Qt.Key_Control,
+            "alt": QtCore.Qt.Key_Alt,
+            "placeholder": None,
+        }
+
+    # ==============================================================================
+    # Providing the basic key and mouse event handling functions.
+    # ==============================================================================
+    def mouse_key_check(self, event, key_status: Dict, mouse_key: Dict):
+        """Giving the mouse key as a dictionary that is in the list of supported keys, return if it is happening (bool).
+
+        Parameters
+        ----------
+        event : PySide2.QtGui.QMouseEvent
+            The Qt event.
+        key_status : Dict
+            The status dictionary that is like `"key_status": { "shift": False, "control": False, "alt": False }`.
+        mouse_key : Dict
+            The dictionary that is like `"mouse_key": { "mouse": "right", "key": "shift" }`.
+
+        Returns
+        ----------
+        bool:
+            If the mouse key is happening.
+
+
+        Notes
+        ----------
+        This function is designed to check mouse_key combine interactions. The interactions are:
+        "box_selection", "selection", "multi_selection", "deselect", "pan", "rotate", for now.
+        """
+        supported_buttons = {
+            "left": QtCore.Qt.LeftButton,
+            "right": QtCore.Qt.RightButton,
+            "middle": QtCore.Qt.MiddleButton,
+        }
+
+        if event.buttons() & supported_buttons[mouse_key["mouse"]]:
+            if key_status.get(mouse_key["key"]) is None:
+                key_status[mouse_key["key"]] = False
+
+            if mouse_key["key"] == "":
+                for key in key_status.values():
+                    if key:
+                        return False
+                return True
+
+            elif key_status[mouse_key["key"]]:
+                return True
+
+        else:
+            return False
+
+    def mouse_check(self, event, button_name: str):
+        """Giving the button name that is one of the supported buttons, return if the button is pressed.
+
+        Parameters
+        ----------
+        event : PySide2.QtGui.QMouseEvent
+            The Qt event.
+        button_name : str
+            The name of the button. It should be exist in the list of `supported buttons`.
+
+        Returns
+        ----------
+        bool:
+            If the button is pressed.
+        """
+        supported_buttons = {
+            "left": QtCore.Qt.LeftButton,
+            "right": QtCore.Qt.RightButton,
+            "middle": QtCore.Qt.MiddleButton,
+        }
+        if button_name not in supported_buttons:
+            # Normally, this should not happen.
+            raise KeyError(f"Button {button_name} is not supported.")
+        else:
+            if event.button() == supported_buttons[button_name]:
+                return True
+            else:
+                return False
+
+    def key_check(self, event, key_status: Dict, key_name: str):
+        """Giving the key name that is one of the supported keys, return if the key is pressed.
+
+        Parameters
+        ----------
+        event : PySide2.QtGui.QMouseEvent
+            The Qt event.
+        key_status : Dict
+            The status dictionary that is like `"key_status": { "shift": False, "control": False, "alt": False }`.
+        key_name : str
+            The name of the key. It should be exist in the list of `supported keys`.
+
+        Returns
+        ----------
+        bool:
+            If the key is pressed.
+        """
+
+        if key_name == "":
+            for key in key_status.values():
+                if key:
+                    return False
+            return True
+        elif key_name not in self.supported_keys:
+            raise KeyError(f"Key {key_name} is not supported.")
+        else:
+            if event.key() == self.supported_keys[key_name]:
+                return True
+            else:
+                return False
 
     def interactive(action="add"):
         """Decorator for transforming functions into "data add" or "object edit" actions.
@@ -240,6 +360,116 @@ class Controller:
         self.app.view.camera.reset_position()
         self.app.view.update_projection()
         self.app.view.update()
+
+    # ==============================================================================
+    # Key mouse actions
+    # ==============================================================================
+
+    def mouse_move_action(self, event):
+        # record mouse position
+        self.mouse.pos = event.pos()
+        # compute displacement
+        dx = self.mouse.dx()
+        dy = self.mouse.dy()
+
+        # * mouse_key
+        # box_selection
+        if self.mouse_key_check(event, self.key_status, self.mouse_key["box_selection"]):
+            self.app.selector.mode = "multi"
+            self.app.selector.perform_box_selection(self.mouse.pos.x(), self.mouse.pos.y())
+        # box_deselection
+        elif self.mouse_key_check(event, self.key_status, self.mouse_key["box_deselection"]):
+            self.app.selector.mode = "deselect"
+            self.app.selector.perform_box_selection(self.mouse.pos.x(), self.mouse.pos.y())
+        # pan
+        elif self.mouse_key_check(event, self.key_status, self.mouse_key["pan"]):
+            self.app.view.camera.pan(dx, dy)
+        # rotate
+        elif self.mouse_key_check(event, self.key_status, self.mouse_key["rotate"]):
+            self.app.view.camera.rotate(dx, dy)
+
+        # record mouse position
+        self.mouse.last_pos = event.pos()
+        self.app.view.update()
+
+    def mouse_press_action(self, event):
+        # * mouse_key
+        # box_selection
+        if self.mouse_key_check(event, self.key_status, self.mouse_key["box_selection"]) or self.mouse_key_check(
+            event, self.key_status, self.mouse_key["box_deselection"]
+        ):
+            self.app.selector.reset_box_selection(event.pos().x(), event.pos().y())
+        # change cursor
+        elif self.mouse_key_check(event, self.key_status, self.mouse_key["pan"]):
+            QtWidgets.QApplication.setOverrideCursor(QtCore.Qt.OpenHandCursor)
+        elif self.mouse_key_check(event, self.key_status, self.mouse_key["rotate"]):
+            QtWidgets.QApplication.setOverrideCursor(QtCore.Qt.SizeAllCursor)
+
+        self.mouse.last_pos = event.pos()
+        self.app.view.update()
+
+    def mouse_release_action(self, event):
+        # * mouse_key
+        # selection
+        if self.mouse_check(event, self.mouse_key["selection"]["mouse"]):
+            if self.app.selector.wait_for_selection_on_plane:
+                self.app.selector.finish_selection_on_plane(event.pos().x(), event.pos().y())
+            else:
+                self.app.selector.enabled = True
+
+        QtWidgets.QApplication.restoreOverrideCursor()
+
+    def wheel_action(self, event):
+        degrees = event.delta() / 8
+        steps = degrees / 15
+        self.app.view.camera.zoom(steps)
+        self.app.view.update()
+
+    def key_press_action(self, event):
+        # * mouse_key
+        # selection
+        if event.key() == QtCore.Qt.Key_Return or event.key() == QtCore.Qt.Key_Enter:
+            self.app.selector.finish_selection()
+            self.key_status[self.mouse_key["selection"]["key"]] = True
+        # multi_selection
+        if self.key_check(event, self.key_status, self.mouse_key["selection"]["multi_selection"]):
+            self.app.selector.mode = "multi"
+            self.key_status[self.mouse_key["selection"]["multi_selection"]] = True
+        # deselect
+        if self.key_check(event, self.key_status, self.mouse_key["selection"]["deselect"]):
+            self.app.selector.mode = "deselect"
+            self.key_status[self.mouse_key["selection"]["deselect"]] = True
+
+        # * key actions
+        for action in self.actions.values():
+            if action.keys_pressed_check(event):
+                action.keys_pressed_action()
+
+        self.app.view.update()
+
+    def key_release_action(self, event):
+        # * box_selection
+        # box_selection
+        if self.key_check(event, self.key_status, self.mouse_key["box_selection"]["key"]):
+            self.key_status[self.mouse_key["box_selection"]["key"]] = False
+        # box_deselection
+        elif self.key_check(event, self.key_status, self.mouse_key["box_deselection"]["key"]):
+            self.key_status[self.mouse_key["box_deselection"]["key"]] = False
+        # pan
+        elif self.key_check(event, self.key_status, self.mouse_key["pan"]["key"]):
+            self.key_status[self.mouse_key["pan"]["key"]] = False
+        # rotate
+        elif self.key_check(event, self.key_status, self.mouse_key["rotate"]["key"]):
+            self.key_status[self.mouse_key["rotate"]["key"]] = False
+
+        # * multi_selection
+        if self.key_check(event, self.key_status, self.mouse_key["selection"]["multi_selection"]):
+            self.app.selector.mode = self.app.selector.overwrite_mode or "single"
+            self.key_status[self.mouse_key["selection"]["multi_selection"]] = False
+        # * deselect
+        if self.key_check(event, self.key_status, self.mouse_key["selection"]["deselect"]):
+            self.app.selector.mode = self.app.selector.overwrite_mode or "single"
+            self.key_status[self.mouse_key["selection"]["deselect"]] = False
 
     # ==============================================================================
     # Scene actions